Optimizes bubbleSort
Previously, the bubbleSort function would keep going through element of the array, disregarding the 'bubbled', sorted number(s) at the end of the list for each pass through. It would sort correctly, but would do this even if the passed in array was already sorted. This new bubbleSort function takes care of that with a while loop, which will only run once if the passed in array is sorted.